requirements.txt
attrs==19.3.0
Automat==20.2.0
configparser==5.0.0
constantly==15.1.0
coverage==5.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
PyHamcrest==2.0.2
six==1.14.0
Twisted==20.3.0
zope.interface==5.0.1
